// packages/toolkit/src/codewhisperer/client/codewhispereruserclient.d.ts
  export type CustomizationArn = string;
// ...
  export interface FileContext {
    leftFileContent: FileContextLeftFileContentString;
    rightFileContent: FileContextRightFileContentString;
    filename: FileContextFilenameString;
    programmingLanguage: ProgrammingLanguage;
  }
  export type FileContextFilenameString = string;
  export type FileContextLeftFileContentString = string;
  export type FileContextRightFileContentString = string;
  export interface FollowupPrompt {
    content: FollowupPromptContentString;
    userIntent?: UserIntent;
  }
  export type FollowupPromptContentString = string;
  export interface GenerateCompletionsRequest {
    fileContext: FileContext;
    maxResults?: GenerateCompletionsRequestMaxResultsInteger;
    nextToken?: GenerateCompletionsRequestNextTokenString;
    referenceTrackerConfiguration?: ReferenceTrackerConfiguration;
    supplementalContexts?: SupplementalContextList;
    customizationArn?: CustomizationArn;
    optOutPreference?: OptOutPreference;
    userContext?: UserContext;
  }
  export type GenerateCompletionsRequestMaxResultsInteger = number;
  export type GenerateCompletionsRequestNextTokenString = string;
// ...
  export type IdeCategory = "JETBRAINS"|"VSCODE"|string;
// ...
  export type OperatingSystem = "MAC"|"WINDOWS"|"LINUX"|string;
  export type OptOutPreference = "OPTIN"|"OPTOUT"|string;
  export type PaginationToken = string;
  export interface Position {
    line: Integer;
    character: Integer;
  }
  export type PreSignedUrl = string;
  export type PrimitiveInteger = number;
  export interface ProgrammingLanguage {
    languageName: ProgrammingLanguageLanguageNameString;
  }
  export type ProgrammingLanguageLanguageNameString = string;
  export type ProgressUpdates = TransformationProgressUpdate[];
  export interface Range {
    start: Position;
    end: Position;
  }
  export type RecommendationsWithReferencesPreference = "BLOCK"|"ALLOW"|string;
  export interface Reference {
    licenseName?: ReferenceLicenseNameString;
    repository?: ReferenceRepositoryString;
    url?: ReferenceUrlString;
    recommendationContentSpan?: Span;
  }
  export type ReferenceLicenseNameString = string;
  export type ReferenceRepositoryString = string;
  export interface ReferenceTrackerConfiguration {
    recommendationsWithReferences: RecommendationsWithReferencesPreference;
  }
// ...
  export type String = string;
  export type SuggestionState = "ACCEPT"|"REJECT"|"DISCARD"|"EMPTY"|string;
  export interface SupplementalContext {
    filePath: SupplementalContextFilePathString;
    content: SupplementalContextContentString;
  }
  export type SupplementalContextContentString = string;
  export type SupplementalContextFilePathString = string;
  export type SupplementalContextList = SupplementalContext[];
// ...
  export type UUID = string;
  export interface UploadContext {
    taskAssistPlanningUploadContext?: TaskAssistPlanningUploadContext;
  }
  export type UploadId = string;
  export type UploadIntent = "TRANSFORMATION"|"TASK_ASSIST_PLANNING"|string;
  export interface UserContext {
    ideCategory: IdeCategory;
    operatingSystem: OperatingSystem;
    product: UserContextProductString;
    clientId?: UUID;
    ideVersion?: String;
  }
  export type UserContextProductString = string;